Fourth Commit

+++ HotelBookingServer/controllers/clerkWebHooks.js
@@ -1,6 +1,6 @@
 const User = require("../models/User");
 const { Webhook } = require("svix");
- 
+
 const clerkWebHooks = async (req, res) => {
   try {
     const wh = new Webhook(process.env.CLERK_WEBHOOK_SECRET);
@@ -11,44 +11,28 @@ const clerkWebHooks = async (req, res) => {
       "svix-signature": req.headers["svix-signature"],
     };
 
-    // Verify signature using RAW BODY BUFFER
-    //const rawBody = req.body; // Buffer
-    const evt= await wh.verify(req.body,headers);
-
-    // Convert to JSON
-    // const bodyString = rawBody.toString("utf8");
-    const { data, type } = evt
+    const evt = wh.verify(req.body, headers); // req.body is RAW
+    const { data, type } = evt;
 
     const userData = {
       _id: data.id,
-      username: data.first_name + " " + data.last_name + "" ,
+      username: `${data.first_name || ""} ${data.last_name || ""}`,
       email: data.email_addresses?.[0]?.email_address || "",
       image: data.image_url || "",
     };
 
-    switch (type) {
-      case "user.created":{
-        await User.create(userData);
-        break;
-      }
-
-      case "user.updated":{
-        await User.findOneAndUpdate({_id:data.id}, userData);
-        break;
-      }
-      case "user.deleted":{
-        await User.findOneAndDelete({_id:data.id});
-        break;
-        }
-      default:
-        break;  
+    if (type === "user.created") {
+      await User.create(userData);
+    } else if (type === "user.updated") {
+      await User.findOneAndUpdate({ _id: data.id }, userData);
+    } else if (type === "user.deleted") {
+      await User.findOneAndDelete({ _id: data.id });
     }
 
-    res.json({ success: true, message: "Webhook received" });
-
+    res.status(200).send({ success: true });
   } catch (err) {
-    console.log("Clerk Webhook Error:", err);
-    res.status(400).json({ success: false, message: err.message });
+    console.error("Webhook Error:", err);
+    res.status(400).send({ error: err.message });
   }
 };
